add unit test for the user api

+++ assessment/src/main/java/com/terminusgroup/assessment/controllers/UserController.java
@@ -3,6 +3,8 @@ package com.terminusgroup.assessment.controllers;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.crossstore.ChangeSetPersister;
+import org.springframework.data.crossstore.ChangeSetPersister.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,11 +19,13 @@ public class UserController {
 	private UserService userService;
 	
 	@GetMapping
+	@ResponseStatus(HttpStatus.OK)
 	public List<UserDto> getUsers() {
 		return userService.getAll();
 	}
 	
 	@GetMapping("/{id}")
+	@ResponseStatus(HttpStatus.OK)
 	public UserDto getUserById(@PathVariable("id") Long id) {
 		return userService.getById(id);
 	}
@@ -34,12 +38,17 @@ public class UserController {
 	
 	@PutMapping
 	@ResponseStatus(HttpStatus.OK)
-	public UserDto updateUser(@RequestBody UserDto user) {
+	public UserDto updateUser(@RequestBody UserDto user) throws NotFoundException {
+		if (user == null || user.getId() == null)
+			throw new NotFoundException();
 		return userService.update(user);
 	}
 	
 	@DeleteMapping("/{id}")
-	public void deleteUser(@PathVariable("id") Long id) {
+	@ResponseStatus(HttpStatus.OK)
+	public void deleteUser(@PathVariable("id") Long id) throws NotFoundException{
+		if (userService.getById(id) == null)
+			throw new NotFoundException();
 		userService.delete(id);
 	}
